This new branch features; -simplified local JSON file -restarting on helper componentry meaning we got rid of how skills and about are displayed, in pursuit of a more modularized, testable, and reusable set of components. -maybe u want to use an advanced style sheet? scss, sass? -organized the file structure so there is 1. pages (the routes for website) 2. helperComponents (what wil be working on) 3.BenAwad.PracticalReactComponents
