+++ src/portfolio.json
@@ -1,5 +1,9 @@
 {
-  "skills" :
+  "skills" : 
+  
+  
+  
+  
   {
     "development" :
     [
